Greatly simplified error handling
Use panics to bail out of the goroutine entirely if we hit an error
that we can't recover from (which there are a lot of), including simple
things such as user disconnects.

Also added a function to generate a simple map, and give the user a
default room to be in if they don't have one.