Added DbObject type and Identifiable interface
Made it so that fields that are common across all database types are contained
in the DbObject type (and inlined). Makes is that we can define high-level
functions like GetId instead of implementing it separately for each type that
gets added.

This also simplifies the commitObject function signature since it can just ask
the object for its ID instead of forcing the caller to pass it in separately.