Significant rework of the Telnet code to make telnet protocol processing seemless
Added a Telnet type that implements net.Conn and can wrap a connection. It strips
all of the telnet codes from the input stream and stores off any subnegotiation parameter
data. Makes it so that the user can interact with the connection object without
worrying about protocol stuff being mixed in with the user input.