Added character.HasItem()
Also moved the check to ensure that we don't set hitpoints > health
into SetHitPoints() rather than Heal()

Also, added a check so that if the max health goes down we also lower
the hitpoint count. It panics due to the poor way the database objects
are managed at the moment.