Tweaked the way database objects are constructed/initialized
The modified() call needs to be in the specific class that is being
created, not in the DbObject, so that we commit all of its fields in
the database, not just the ones in DbObject.

Also made it so we bypass the setter methods for initializing data
so that the entire object is initialized before we hand it off to
be committed to the database.