Improvements to event handling and saving objects to the DB
Fixes a deadlock issue whenever the channel that receives modified
object IDs filled up. We now save all modified objects on a periodic
timer rather than synchronously with every change to an object field.

Updated the event loop to check for whether or not the receiver of
an event actually cares about it before sending it along the channel.

Also made it so we crash if a session gets too far behind in processing
its events.