Fixed a bug in the telnet Read() implementation
Noticed this issue when I resized the terminal in rapid succession for a while
and it caused the connection to drop due to an io error.

Turns out that the bufio scanner will freak out if it has more than 100
consecutive reads that return with no data and no error. This change
makes it so that we don't return from Read() unless there's actually something
to return.